Compile directory
I was under the impression that Babel would bundle source dependencies but that is not the case.
